requirements.txt
python-telegram-bot==20.0a0
python-dotenv
-r ./src/scrap_diputades/requirements.txt